Revamped the entire process of how items are deleted from the database. The new process is a lot more object oriented with a lot more safeguard checking. This now completes the delete album feature, where an album no longer has to be empty to be deleted.
Also implemented the delete artist feature and cleaned up some other delete
functions. The delete artist link is only shown when an artist has no more
albums.

+++ cgi-bin/Artist.pm
@@ -104,12 +104,25 @@ sub Insert
 sub Remove
 {
     my ($this) = @_;
-    my ($sql);
+    my ($sql, $refcount);
 
     return if (!defined $this->GetId());
-  
+
     $sql = Sql->new($this->{DBH});
+
+    # See if there are any tracks that needs this artist
+    ($refcount) = $sql->GetSingleRow("Track", ["count(*)"],
+                                   [ "Track.artist", $this->GetId()]);
+    return undef if ($refcount > 0);
+  
+    # See if there are any albums that needs this artist
+    ($refcount) = $sql->GetSingleRow("Album", ["count(*)"],
+                                   [ "Album.artist", $this->GetId()]);
+    return undef if ($refcount > 0);
+  
     $sql->Do("delete from Artist where id = " . $this->GetId());
+
+    return 1;
 }
 
 # Load an artist record given a name. The name must match exactly.
